+++ src/shared/components/button/ctaButton/CtaButton.tsx
@@ -1,5 +1,5 @@
 import KakaoIcon from '@assets/icons/kakaoIcon.svg?react';
-import * as styles from './ctaButton.css';
+import * as styles from './CtaButton.css';
 
 interface CtaButtonProps extends React.ComponentProps<'button'> {
   children: React.ReactNode;
